backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# Supabase (optional, for direct Supabase client usage)
supabase==2.7.0

# AI / Image Analysis
# detectron2 requires separate installation (see README)
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
torch==2.4.0
torchvision==0.19.0

# Background tasks
apscheduler==3.10.4

# Weather
aiohttp==3.10.0

# Export
openpyxl==3.1.5

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Dev
python-dotenv==1.0.1
python-multipart==0.0.9
